Data generating process: linear transformation model with non-compliance

TODO: ADD MORE DESCRIPTION HERE.

dgp_internal(
  n = 200,
  x1_dist = "normal",
  x2_dist = "normal",
  sigma_x = 1,
  nonlinearity = "sin",
  ps_link = "logit",
  ps_trim = 0.05,
  eta_0 = 0,
  eta_1 = 0,
  eta_2 = 0,
  eta_3 = 0,
  eta_4 = 0,
  p_c = 0.7,
  rho_a = 0.5,
  nc_scale = 0.5,
  nc_shift = 1,
  alpha = 1,
  beta_1 = 0.5,
  beta_2 = -0.7,
  beta_3 = 0,
  sigma_e = 1,
  error_dist = "gumbel"
)

Arguments

n

numeric Number of samples (observations).

x1_dist

character "binary", "uniform", "normal".

x2_dist

character "binary", "uniform", "normal".

sigma_x

numeric Standard deviation of x (if not binary).

nonlinearity

character Apply what nonlinear transformation to x?

ps_link

character link function for propensity score DGP, see ?make.link.

ps_trim

numeric This enforces positivity assumption. (ps_trim, 1 - ps_trim) will be the bound of generated propensity score, implemented by rejection sampling.

eta_0

numeric Coefficients for the propensity score DGP.

eta_1

numeric Coefficients for the propensity score DGP.

eta_2

numeric Coefficients for the propensity score DGP.

eta_3

numeric Coefficients for the propensity score DGP.

eta_4

numeric Coefficients for the propensity score DGP.

p_c

numeric Compliance rate.

rho_a

numeric Proportion of always-takers among non-compliers.

nc_scale

numeric Scale non-complier coefficients.

nc_shift

numeric Shift non-complier outcomes.

alpha

numeric Coefficient associated with the treatment difference

beta_1

numeric Coefficients associated with x1.

beta_2

numeric Coefficients associated with x2.

beta_3

numeric Coefficients associated with x1 * x2.

sigma_e

standard deviation of the error term, default is 1.

error_dist

"gumbel" or "normal".

Value

df - generated data frame with columns x1, x2, x2_t, lp, ps, z, c, a, ey, y
df$... - see "dgp_prop_score()"
df$z - randomized treatment assignment based on the propensity scores
df$c - complier status 0, 1, 2 = always-taker, complier, never-taker
df$a - actual treatment received
df$ey - expected value of y
df$y - outcome